# server/openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: PlanTogether Server
  version: 0.1.0
  description: |
    Launch MVP 서버 API. 기준 설계는 docs/account_backend_design.md다.

    인증 endpoint(/v1/auth/*)는 §6의 Idempotency-Key 규칙 밖이다. 로그인
    시점에는 idempotency 키를 묶을 user_id·device_id가 없다.
security:
  - bearerAuth: []
servers:
  - url: http://localhost:8080
    description: 로컬 개발
paths:
...
  /v1/auth/apple:
    post:
      summary: Apple 로그인·가입
      description: |
        account_backend_design.md §5.1. identity token의 서명·iss·aud·exp·nonce를
        검증하고 authorization code를 즉시 교환한다. 같은 Apple subject는 같은
        user_id에 연결된다. 로그인할 때마다 서버가 새 device_id를 발급한다.

        raw_nonce는 iOS가 만든 원문이다. iOS는 그 SHA-256 hex를 Apple 요청의
        nonce로 보내야 한다.
      operationId: signInWithApple
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
      responses:
        "200":
          description: 로그인 성공.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "423":
          $ref: "#/components/responses/AccountLocked"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/auth/refresh:
    post:
      summary: refresh token 회전
      description: |
        §5.2. 성공하면 이전 refresh token은 폐기되고 새 쌍이 발급된다. 이미 쓰인
        token을 다시 보내면 탈취로 간주해 그 기기의 token family와 기기를 모두
        폐기하고 401을 반환한다. 응답을 잃고 같은 token으로 재시도해도 같다.
      operationId: refreshSession
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: 회전 성공.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "423":
          $ref: "#/components/responses/AccountLocked"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/logout:
    post:
      summary: 현재 기기 로그아웃
      description: |
        §5.2. 현재 기기의 세션과 기기 자체를 폐기한다. 다른 기기는 영향이 없다.
        이미 로그아웃한 기기에 다시 호출해도 성공한다(access token이 아직
        인증되는 동안).
      operationId: logout
      tags: [auth]
      responses:
        "204":
          description: 로그아웃됐다.
        "401":
          $ref: "#/components/responses/InvalidSession"
        "423":
          $ref: "#/components/responses/AccountLocked"
...
components:
...
  headers:
    RetryAfter:
      description: 재시도까지 기다릴 초(정수, 최소 1). 429에서는 창 종료까지 남은 초다. 남은 횟수와 한도는 알리지 않는다.
      schema: {type: integer, minimum: 1}
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 15분 수명의 access token. 서명 외에 기기 폐기와 계정 상태를 매 요청 DB에서 확인한다.
  responses:
...
    InvalidRequest:
      description: 형식 또는 필드 검증 실패.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidSession:
      description: 자격이 무효다. 원인(만료, 폐기, 위조, 재사용)을 구분하지 않는다. 클라이언트는 refresh를 시도하고, refresh도 401이면 다시 로그인한다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AccountLocked:
      description: 계정이 active가 아니다(disabled, deletion_requested, deleting, deleted).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: |
        요청 제한 초과(docs/rate_limit_design.md §6). Retry-After 이후에 재시도한다. 본문은 어느 키가
        초과했는지, 한도, 남은 횟수를 담지 않는다. **세션 상실이 아니다.** 클라이언트는 로그아웃하거나
        refresh token을 폐기하지 않는다.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unavailable:
      description: Apple 같은 외부 의존성 장애, 또는 요청 제한 장치를 확인하지 못한 경우(closed scope). 사용자 자격 문제가 아니므로 클라이언트는 로그아웃하지 않고 Retry-After 이후 재시도한다.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [invalid_request, invalid_session, forbidden, not_found, version_conflict,
                 idempotency_mismatch, sync_cursor_expired, account_locked, rate_limited, internal]
        message:
          type: string
          description: 사람이 읽는 짧은 설명. 클라이언트는 code로 분기한다.
        request_id:
          type: string
          format: uuid
        current_version:
          type: integer
          format: int64
          description: version_conflict에서만 있다.
    SignInRequest:
      type: object
      required: [identity_token, authorization_code, raw_nonce]
      additionalProperties: false
      properties:
        identity_token:
          type: string
        authorization_code:
          type: string
          description: 1회용, 5분 유효. 서버가 즉시 Apple과 교환한다.
        raw_nonce:
          type: string
        display_name:
          type: string
          description: Apple이 최초 동의 때 준 이름. 신규 가입에만 쓴다. 제어 문자는 지우고 50자로 자른다.
    RefreshRequest:
      type: object
      required: [refresh_token]
      additionalProperties: false
      properties:
        refresh_token:
          type: string
    SessionResponse:
      type: object
      required: [request_id, user_id, device_id, access_token, access_token_expires_at,
                 refresh_token, refresh_token_expires_at, new_user]
      additionalProperties: false
      properties:
        request_id: {type: string, format: uuid}
        user_id: {type: string, format: uuid}
        device_id:
          type: string
          format: uuid
          description: 서버가 발급한 기기 ID. access token에 결합돼 있다.
        access_token: {type: string}
        access_token_expires_at: {type: string, format: date-time}
        refresh_token:
          type: string
          description: Keychain에만 저장한다. 로그·분석 이벤트에 남기지 않는다.
        refresh_token_expires_at: {type: string, format: date-time}
        new_user: {type: boolean}
...